+++ src/main/java/net/webpdf/wsclient/session/RestSession.java
@@ -6,18 +6,21 @@ import net.webpdf.wsclient.exception.ResultException;
 import net.webpdf.wsclient.http.HttpMethod;
 import net.webpdf.wsclient.http.HttpRestRequest;
 import net.webpdf.wsclient.https.TLSContext;
+import net.webpdf.wsclient.proxy.ProxyConfiguration;
 import net.webpdf.wsclient.schema.beans.Token;
 import net.webpdf.wsclient.schema.beans.UserCredentialsBean;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.client.HttpClients;
+import org.apache.http.impl.conn.DefaultProxyRoutePlanner;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
 import java.net.URL;
 
+@SuppressWarnings("unused")
 public class RestSession extends AbstractSession {
 
     private static final String INFO_PATH = "authentication/user/info/";
@@ -27,10 +30,12 @@ public class RestSession extends AbstractSession {
     private Token token = new Token();
     @Nullable
     private UserCredentialsBean userCredentials = new UserCredentialsBean();
-    @NotNull
+    @Nullable
     private CloseableHttpClient httpClient;
     @NotNull
     private DocumentManager documentManager = new DocumentManager(this);
+    @NotNull
+    private final HttpClientBuilder httpClientBuilder;
 
     /**
      * Creates new {@link RestSession} instance
@@ -44,13 +49,12 @@ public class RestSession extends AbstractSession {
         this.dataFormat = DataFormat.JSON;
 
         RequestConfig clientConfig = RequestConfig.custom().setAuthenticationEnabled(true).build();
-        HttpClientBuilder httpClientBuilder = HttpClients.custom()
-                                                  .setDefaultRequestConfig(clientConfig)
-                                                  .setDefaultCredentialsProvider(this.credentialsProvider);
+        this.httpClientBuilder = HttpClients.custom()
+                                     .setDefaultRequestConfig(clientConfig)
+                                     .setDefaultCredentialsProvider(this.credentialsProvider);
         if (getTlsContext() != null) {
             httpClientBuilder.setSSLContext(getTlsContext().getSslContext());
         }
-        httpClient = httpClientBuilder.build();
     }
 
     /**
@@ -64,13 +68,27 @@ public class RestSession extends AbstractSession {
     }
 
     /**
+     * Set a {@link ProxyConfiguration}.
+     *
+     * @param proxy The {@link ProxyConfiguration}, that shall be set.
+     * @throws ResultException Shall be thrown, when resolving the proxy failed.
+     */
+    @Override
+    public void setProxy(@Nullable ProxyConfiguration proxy) throws ResultException {
+        super.setProxy(proxy);
+        if (proxy != null) {
+            httpClientBuilder.setRoutePlanner(new DefaultProxyRoutePlanner(proxy.getHost()));
+        }
+    }
+
+    /**
      * Returns the {@link CloseableHttpClient} connected to the server via this session.
      *
      * @return The {@link CloseableHttpClient} connected to the server via this session.
      */
     @NotNull
     public CloseableHttpClient getHttpClient() {
-        return this.httpClient;
+        return this.httpClient == null ? this.httpClient = this.httpClientBuilder.build() : this.httpClient;
     }
 
     /**
@@ -106,7 +124,9 @@ public class RestSession extends AbstractSession {
         } catch (ResultException ex) {
             throw new IOException("Unable to logout from server", ex);
         } finally {
-            this.httpClient.close();
+            if (this.httpClient != null) {
+                this.httpClient.close();
+            }
         }
     }